fix: numerous memory-related bugs; update: `README.md` information; improve: ELF utils code; add: `CMakeLists.txt`
This commit improves the overall repository compatibility, stability and performance. It fixes numerous bugs related to memory that previously lead to crashes and stability issues, updates the information in "README.md" file, improves the ELF utils code to enhance stability but mainly not conflict with other ELF utils, and adds cmake-based compiling.